Add remove confirmation

+++ js/app.controller.js
@@ -69,6 +69,13 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+
+    // ex 1 
+    //Remove location – add confirmation (use confirm)
+
+    const isConfirm = confirm('Are you sure you want to remove this location?')
+    if (!isConfirm) return
+
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
@@ -223,7 +230,7 @@ function getFilterByFromQueryParams() {
     const queryParams = new URLSearchParams(window.location.search)
     const txt = queryParams.get('txt') || ''
     const minRate = queryParams.get('minRate') || 0
-    locService.setFilterBy({txt, minRate})
+    locService.setFilterBy({ txt, minRate })
 
     document.querySelector('input[name="filter-by-txt"]').value = txt
     document.querySelector('input[name="filter-by-rate"]').value = minRate
